+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
 # MixinMCP Changelog
+## [0.6.3]
+- Fix unresolvable dependencies preventing full gradle sync due to decompile task
+- Show pop-up warning when some dependencies were not decompiled
+
 ## [0.6.2]
 - Auto gitignore injected rules